requirements.txt
numpy<2.0
scipy
librosa==0.10.2
numba

torch
torchaudio

faster-whisper
ctranslate2>=4.0,<5

sounddevice
soundfile
requests

openai>=2

transformers>=4.43
sentencepiece
tokenizers>=0.13,<1
huggingface_hub>=0.13

onnxruntime
funasr==1.0.27
modelscope==1.10.0

g2p_en
jieba
fast_langdetect>=0.3.1

PyYAML
psutil
chardet

av>=11
tqdm
ffmpeg-python
